fix build for OpenRM variant
nvidia-x.y/nvidia-peermem/ contains a copy of nv-p2p.h, so it may get picked as NVIDIA_SRC_DIR.
Later nv-frontend.c is searched there and cannot be found because it is located in a different subdir.
The fix is to search for nv-p2p.c, which is colocated along with nv-frontend.c.